[inference] moved ops tests to test_infer
