+++ tests/test_infer/test_ops/triton/test_xine_copy.py
